refactor: Return List<Dvd> instead of Dvd[]
Found better solution for client app's issue with marshalling JSON to a parameterized collection of objects (type erasure when deserializing).

Previous solution was to use an array-that-is-not-big-enough-to-fit-returned-data in order to have a new array of the-same-type be created under the hood.

Now the client app no longer uses RestTemplate's getForObject(). Instead, it uses exchange() which passes a ParameterizedTypeReference as an argument.